requirements.txt
flask
google-generativeai
python-dotenv
gtts
protobuf==4.25.3
